+++ library_generation/test/integration_tests.py
@@ -26,6 +26,7 @@ from typing import Dict
 from library_generation.generate_repo import generate_from_yaml
 from library_generation.model.generation_config import from_yaml
 from library_generation.test.compare_poms import compare_xml
+from library_generation.utilities import get_commit_messages
 from library_generation.utilities import (
     get_library_name,
     sh_util as shell_call,
@@ -45,6 +46,21 @@ output_folder = shell_call("get_output_folder")
 
 
 class IntegrationTest(unittest.TestCase):
+    def test_get_commit_message_success(self):
+        repo_url = "https://github.com/googleapis/googleapis.git"
+        new_committish = "5c5ecf0eb9bd8acf5bed57fe1ce0df1770466089"
+        old_committish = "e04130efabc98c20c56675b0c0af603087681f48"
+        message = get_commit_messages(
+            repo_url=repo_url,
+            new_committish=new_committish,
+            old_committish=old_committish,
+        )
+        self.assertTrue("5c5ecf0eb9bd8acf5bed57fe1ce0df1770466089" in message)
+        self.assertTrue("8df1cd698e2fe0b7d1c298dabafdf13aa01c4d39" in message)
+        self.assertTrue("87a71a910ce60009cc578eb183062fb7d62e664f" in message)
+        self.assertTrue("cfda64661b5e005730e9c7f24745ff2e40680647" in message)
+        self.assertFalse("e04130efabc98c20c56675b0c0af603087681f48" in message)
+
     def test_generate_repo(self):
         shutil.rmtree(f"{golden_dir}", ignore_errors=True)
         os.makedirs(f"{golden_dir}", exist_ok=True)
